2003-07-01  Timothy S. Stack <stack@cs.utah.edu>
	* configure.in:
	Autoheader doesn't understand --localdir anymore.

	* kaffe/kaffevm/jit3/basecode.c,
	kaffe/kaffevm/jit3/basecode.h,
	kaffe/kaffevm/jit3/icode.c,
	kaffe/kaffevm/jit3/registers.h:
	Load/Store float constants as floats and not doubles, otherwise
	the jitted code will read floats from the constant pool that were
	stored as doubles.

	Reported by: Gerlando Falauto <iurly@writeme.com>

+++ kaffe/kaffevm/jit3/basecode.h
@@ -119,7 +119,7 @@ void _slot_slot_const(SlotInfo*, SlotInfo*, jword, ifunc, int);
 		void func(sequence*);					\
 		_slot_slot_fconst((dst), (src), (src2), (func), (t));	\
 	}
-void _slot_slot_fconst(SlotInfo*, SlotInfo*, double, ifunc, int);
+void _slot_slot_fconst(SlotInfo*, SlotInfo*, float, ifunc, int);
 
 #define	slot_slot_slot(dst, src, src2, func, t)				\
 	{								\
